refactor: group runtime utilities and gateway modules
