+++ docs/CAPABILITY_MATRIX.md
@@ -10,12 +10,12 @@ This is the authoritative record of what is real versus planned.
 | Causal memory short-circuit cache | Implemented | reasoning/causal_memory.py, reasoning/elevator.py, state/store.py | Optional `reasoning.causal_memory.enabled` path short-circuits repeated trigger patterns with cached operator-safe resolutions, reducing LLM variance and token spend while preserving trigger-tier authority. |
 | Prompt history placeholder interpolation | Implemented | reasoning/elevator.py, tests/test_elevator.py | Supports safe `{history.method(...)}` interpolation in prompts (AST-validated, async-resolved), including compact `history.last_n(...)` output for low-token local SLM reasoning. Unresolved placeholders deterministically substitute `null` sentinel rather than passing raw template tokens to the model. |
 | Skill prompt history placeholder guard | Implemented | skills/loader.py, tests/test_skill_loader.py | Skill load now fails fast when any prompt template exceeds 16 `{history.*}` placeholders; unresolved/unsupported placeholders are emitted at WARNING level for operator visibility. |
-| Intelligence Elevator Tier 3 (gateway) | Implemented | elevator.py, reasoning/escalation_policy.py, gateway_reasoning.py | Deterministic escalation signals are evaluated before local SLM inference and can route to gateway over MQTT (`ori/{device_id}/reasoning/request` → `ori/{device_id}/reasoning/response`). Non-explicit signals fall back to local_slm when transport is unavailable; explicit `escalate_to: gateway` returns a gateway-unavailable stub rather than silently downgrading. Cloud reasoning, when used, is a gateway backend, not a runtime dependency. |
+| Intelligence Elevator Tier 3 (gateway) | Implemented | elevator.py, reasoning/escalation_policy.py, gateway/reasoning.py | Deterministic escalation signals are evaluated before local SLM inference and can route to gateway over MQTT (`ori/{device_id}/reasoning/request` → `ori/{device_id}/reasoning/response`). Non-explicit signals fall back to local_slm when transport is unavailable; explicit `escalate_to: gateway` returns a gateway-unavailable stub rather than silently downgrading. Cloud reasoning, when used, is a gateway backend, not a runtime dependency. |
 | CoAP actuation executor | Implemented | actions/coap.py, runtime.py:312 | Host allowlist enforced in config.py:409. |
 | CoAP HAL telemetry adapter | Implemented | hal/coap_adapter.py, runtime.py, config.py | Poll-loop + cached-read CoAP telemetry path with circuit breaker, aiocoap context lifecycle shutdown, and `actions.coap.allowed_hosts` boundary enforcement for sensor targets. |
 | Tier C approval workflow | Implemented | action_dispatcher.py:649 | Uses unified alert failover sender (primary + secondary channel) with local-console fallback path when comms are unavailable. Approval requests include a short `proposal_id`; scoped `YES-<id>`/`NO-<id>` replies must match the active proposal, and the proposal ID is persisted in action/Tier C decision logs. |
 | Tier C decision logging | Implemented | reasoning/elevator.py, reasoning/action_dispatcher.py, runtime.py, state/store.py | Persists rich approval/rejection/timeout records with sensor context, bounded history window, runtime site metadata, skill/trigger metadata, reasoning confidence/model, operator decision latency, safe-default usage, and final action result for future approval-pattern learning. Bounded export supports `device_id`, `since_ms`, `until_ms`, and `limit` filters for product/cloud sync. |
-| Gateway/reporting data exports | Implemented | state/store.py, gateway_export.py, runtime.py | Runtime exposes provider-neutral bounded exports for action logs, Tier C decision logs, reasoning logs, sensor history, and health. Gateway/product services request data over MQTT (`ori/{device_id}/export/request` → `ori/{device_id}/export/response/{request_id}`) without direct SQLite table knowledge or runtime cloud SDK dependencies. Sensor-history responses support `bucket_ms` aggregation and pagination for weekly reports. Reasoning-log export includes `reasoning_status` and `correlation_id` for reliable action/reasoning joins. |
+| Gateway/reporting data exports | Implemented | state/store.py, gateway/export.py, runtime.py | Runtime exposes provider-neutral bounded exports for action logs, Tier C decision logs, reasoning logs, sensor history, and health. Gateway/product services request data over MQTT (`ori/{device_id}/export/request` → `ori/{device_id}/export/response/{request_id}`) without direct SQLite table knowledge or runtime cloud SDK dependencies. Sensor-history responses support `bucket_ms` aggregation and pagination for weekly reports. Reasoning-log export includes `reasoning_status` and `correlation_id` for reliable action/reasoning joins. |
 | Tier C local console fallback | Implemented | action_dispatcher.py, runtime.py, config.py | When SMS/WhatsApp are unavailable and `actions.local_console.enabled=true`, Tier C approval polls local inbound channel and records `LOCAL:<reply>` in action results. Structured remote command payloads on this channel are consumed, durably audited, and ignored as approval replies; local console is not an unauthenticated remote command ingress. Unrecognised local input is logged and ignored until timeout. |
 | Tier D relay bypass when DevicePolicy restricted | Implemented | action_dispatcher.py | P3-R0b-i. Critical safety fix. |
 | SMS via Africa's Talking (IP) | Implemented | sms.py:57, runtime.py | send(message, to_number). Requires internet. Inbound SMS webhook path defaults to localhost bind (`127.0.0.1`), requires shared-token auth, and enforces 64KB request-body cap to prevent memory exhaustion. |
