+++ ori/actions/sms.py
@@ -29,7 +29,6 @@ import time
 from collections.abc import Awaitable, Callable
 from typing import Any
 
-from ori.bool_utils import is_truthy
 from ori.security.remote_command_responses import (
     format_remote_command_execution_response,
     format_remote_command_rejection_response,
@@ -43,7 +42,8 @@ from ori.security.remote_commands import (
     RemoteCommandVerifier,
     verify_inbound_remote_command,
 )
-from ori.time_utils import now_ms
+from ori.utils.bool_utils import is_truthy
+from ori.utils.time_utils import now_ms
 
 logger = logging.getLogger(__name__)
