+++ ori/hal/growatt_adapter.py
@@ -12,7 +12,7 @@ from ori.hal.base import (
     HardwareCircuitBreaker,
 )
 from ori.network.events import SensorReading
-from ori.time_utils import now_ms
+from ori.utils.time_utils import now_ms
 
 logger = logging.getLogger(__name__)
