+++ ori/hal/mqtt_base.py
@@ -14,7 +14,7 @@ from ori.hal.base import (
     BaseAdapter,
     HardwareCircuitBreaker,
 )
-from ori.time_utils import now_ms
+from ori.utils.time_utils import now_ms
 
 logger = logging.getLogger(__name__)
