+++ ori/reasoning/rule_engine.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass
 from typing import Any
 
 from ori.network.events import OriEvent
-from ori.time_utils import now_ms
+from ori.utils.time_utils import now_ms
 
 logger = logging.getLogger(__name__)
