+++ ori/utils/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2026 Ori Nexus Systems LTD
+# SPDX-License-Identifier: Apache-2.0
